Fix problem where fseq file data gets dumped on load due to number of channels not being a multiple of 4

+++ xSchedule/ScheduleManager.cpp
@@ -661,7 +661,7 @@ int ScheduleManager::Frame(bool outputframe)
 
 void ScheduleManager::CreateBrightnessArray()
 {
-    for (int i = 0; i < 256; i++)
+    for (size_t i = 0; i < 256; i++)
     {
         _brightnessArray[i] = (wxByte)(((i * _brightness) / 100) & 0xFF);
     }
